LPS-53795 move for consistency
